feat(jsx-renderer): Change jsx renderer to be treated as middleware.
